+++ Classes/Domain/Repository/ProcessRepository.php
@@ -32,16 +32,18 @@ use AOE\Crawler\Configuration\ExtensionConfigurationProvider;
 use AOE\Crawler\Domain\Model\Process;
 use AOE\Crawler\Domain\Model\ProcessCollection;
 use Doctrine\DBAL\ArrayParameterType;
+use Doctrine\DBAL\Exception;
 use Symfony\Contracts\Service\Attribute\Required;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Database\Query\QueryBuilder;
+use TYPO3\CMS\Core\Database\Query\Restriction\DeletedRestriction;
+use TYPO3\CMS\Core\Database\Query\Restriction\HiddenRestriction;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Extbase\Persistence\Repository;
 
 /**
  * @internal since v9.2.5
  */
-class ProcessRepository extends Repository
+class ProcessRepository
 {
     final public const TABLE_NAME = 'tx_crawler_process';
 
@@ -66,12 +68,16 @@ class ProcessRepository extends Repository
 
     /**
      * This method is used to find all cli processes within a limit.
+     * @throws Exception
      */
     public function findAll(): ProcessCollection
     {
         /** @var ProcessCollection $collection */
         $collection = GeneralUtility::makeInstance(ProcessCollection::class);
         $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable(self::TABLE_NAME);
+        $queryBuilder->getRestrictions()
+            ->removeByType(HiddenRestriction::class)
+            ->removeByType(DeletedRestriction::class);
 
         $statement = $queryBuilder
             ->select('*')
@@ -142,16 +148,19 @@ class ProcessRepository extends Repository
      */
     public function getActiveProcessesOlderThanOneHour()
     {
+        $processMaxRunTime = $this->extensionSettings['processMaxRunTime'] ?? 360;
+
         $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable(self::TABLE_NAME);
+        $queryBuilder->getRestrictions()
+            ->removeByType(HiddenRestriction::class)
+            ->removeByType(DeletedRestriction::class);
+
         $activeProcesses = [];
         $statement = $queryBuilder
             ->select('process_id', 'system_process_id')
             ->from(self::TABLE_NAME)
             ->where(
-                $queryBuilder->expr()->lte(
-                    'ttl',
-                    intval(time() - $this->extensionSettings['processMaxRunTime'] - 3600)
-                ),
+                $queryBuilder->expr()->lte('ttl', intval(time() - $processMaxRunTime - 3600)),
                 $queryBuilder->expr()->eq('active', 1)
             )
             ->executeQuery();
@@ -172,12 +181,17 @@ class ProcessRepository extends Repository
     public function getActiveOrphanProcesses()
     {
         $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable(self::TABLE_NAME);
+        $queryBuilder->getRestrictions()
+            ->removeByType(HiddenRestriction::class)
+            ->removeByType(DeletedRestriction::class);
+
+        $processMaxRunTime = $this->extensionSettings['processMaxRunTime'] ?? 360;
 
         return $queryBuilder
             ->select('process_id', 'system_process_id')
             ->from(self::TABLE_NAME)
             ->where(
-                $queryBuilder->expr()->lte('ttl', intval(time() - $this->extensionSettings['processMaxRunTime'])),
+                $queryBuilder->expr()->lte('ttl', intval(time() - $processMaxRunTime)),
                 $queryBuilder->expr()->eq('active', 1)
             )
             ->executeQuery()->fetchAllAssociative();
@@ -212,7 +226,7 @@ class ProcessRepository extends Repository
         $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable(self::TABLE_NAME);
         $queryBuilder
             ->delete(self::TABLE_NAME)
-            ->where($queryBuilder->expr()->eq('deleted', 1))
+            ->where($queryBuilder->expr()->eq('deleted', $queryBuilder->createNamedParameter(1)))
             ->executeStatement();
     }