+++ Tests/Functional/Controller/CrawlerControllerTest.php
@@ -26,7 +26,6 @@ use AOE\Crawler\Tests\Functional\BackendRequestTestTrait;
 use AOE\Crawler\Value\QueueFilter;
 use PHPUnit\Framework\MockObject\MockObject;
 use TYPO3\CMS\Core\Authentication\BackendUserAuthentication;
-use TYPO3\CMS\Core\Imaging\IconFactory;
 use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\TestingFramework\Core\AccessibleObjectInterface;
 use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
@@ -56,12 +55,11 @@ class CrawlerControllerTest extends FunctionalTestCase
 
         $mockedQueueRepository = $this->createMock(QueueRepository::class);
         $mockedProcessRepository = $this->createMock(ProcessRepository::class);
-        $mockedIconFactory = $this->createMock(IconFactory::class);
 
         $this->subject = $this->getAccessibleMock(
             CrawlerController::class,
             null,
-            [$mockedQueueRepository, $mockedProcessRepository, $mockedIconFactory]
+            [$mockedQueueRepository, $mockedProcessRepository]
         );
     }